+++ CHANGELOG.md
@@ -4,6 +4,17 @@ seotitle: MarkdownMaster CMS Change History
 description: Compiled list of changes to the MarkdownMaster CMS application and archived download links.
 ---
 
+## MarkdownMaster CMS 5.0.4 - 2025.XX.XX
+
+### New Features
+
+### Changes
+
+### Fixes
+
+* Fix support for newlines within HTML and script tags
+
+
 ## [MarkdownMaster CMS 5.0.3 - 2025.04.27](https://github.com/eVAL-Agency/MarkdownMasterCMS/releases/tag/v5.0.3)
 
 ### New Features